GraphAPI/requirements.txt
azure-storage-blob
msal
opencensus-ext-azure
pytest
requests
structlog
tenacity

--pre
azure-storage-file-datalake